+++ src/Migration/ResourceModel/Destination.php
@@ -16,6 +16,11 @@ class Destination extends AbstractResource
     const CONFIG_DOCUMENT_PREFIX = 'dest_prefix';
 
     /**
+     * @var string
+     */
+    protected $documentPrefix;
+
+    /**
      * Save data into destination resource
      *
      * @param string $documentName
@@ -38,13 +43,13 @@ class Destination extends AbstractResource
                 $data[] = $row;
             }
             if ($i == $pageSize) {
-                $this->adapter->insertRecords($documentName, $data, $updateOnDuplicate);
+                $this->getAdapter()->insertRecords($documentName, $data, $updateOnDuplicate);
                 $data = [];
                 $i = 0;
             }
         }
         if ($i > 0) {
-            $this->adapter->insertRecords($documentName, $data, $updateOnDuplicate);
+            $this->getAdapter()->insertRecords($documentName, $data, $updateOnDuplicate);
         }
         return $this;
     }
@@ -75,7 +80,7 @@ class Destination extends AbstractResource
      */
     public function clearDocument($documentName)
     {
-        $this->adapter->deleteAllRecords($this->addDocumentPrefix($documentName));
+        $this->getAdapter()->deleteAllRecords($this->addDocumentPrefix($documentName));
     }
 
     /**
@@ -83,7 +88,10 @@ class Destination extends AbstractResource
      */
     protected function getDocumentPrefix()
     {
-        return $this->configReader->getOption(self::CONFIG_DOCUMENT_PREFIX);
+        if (null === $this->documentPrefix) {
+            $this->documentPrefix = $this->configReader->getOption(self::CONFIG_DOCUMENT_PREFIX);
+        }
+        return $this->documentPrefix;
     }
 
     /**
@@ -127,7 +135,7 @@ class Destination extends AbstractResource
             $data[] = $row->getData();
         }
         if (!empty($data)) {
-            $this->adapter->updateChangedRecords($this->addDocumentPrefix($documentName), $data);
+            $this->getAdapter()->updateChangedRecords($this->addDocumentPrefix($documentName), $data);
         }
     }
 }