+++ tests/integration/testsuite/Migration/Step/UrlRewrite/Version11410to2000Test.php
@@ -8,7 +8,7 @@ namespace Migration\Step\UrlRewrite;
 
 /**
  * UrlRewrite step test class
- * @dbFixture url_rewrite
+ * @dbFixture url_rewrite_11410
  */
 class Version11410to2000Test extends \PHPUnit_Framework_TestCase
 {