+++ tests/unit/testsuite/Migration/Step/Eav/VolumeTest.php
@@ -59,7 +59,8 @@ class VolumeTest extends \PHPUnit_Framework_TestCase
                     'getDestinationRecords',
                     'getSourceRecordsCount',
                     'getDestinationRecordsCount',
-                    'deleteBackups'
+                    'deleteBackups',
+                    'clearIgnoredAttributes'
                 ]
             )->getMock();
         $this->logger = $this->getMockBuilder('\Migration\Logger\Logger')->disableOriginalConstructor()
@@ -153,6 +154,8 @@ class VolumeTest extends \PHPUnit_Framework_TestCase
         $this->initialData->expects($this->once())->method('getAttributeGroups')->willReturn(1);
         $this->helper->expects($this->any())->method('getDestinationRecordsCount')->willReturn(2);
         $this->helper->expects($this->once())->method('deleteBackups');
+        $this->helper->expects($this->any())->method('clearIgnoredAttributes')->with($eavAttributes)
+            ->willReturn($eavAttributes);
         $this->logger->expects($this->never())->method('addRecord');
 
         $documentsMap = $this->getDocumentsMap();